fixed dumping and uploading annotation if annotation task contains interpolation shapes

+++ cvat/apps/engine/models.py
@@ -139,7 +139,7 @@ class Annotation(models.Model):
     label = models.ForeignKey(Label, on_delete=models.CASCADE)
     frame = models.PositiveIntegerField()
     group_id = models.PositiveIntegerField(default=0)
-    client_id = models.IntegerField(default=-1)
+    client_id = models.BigIntegerField(default=-1)
     class Meta:
         abstract = True
